+++ sdk/appconfiguration/azure-data-appconfiguration/src/main/java/com/azure/data/appconfiguration/ConfigurationAsyncClient.java
@@ -1093,9 +1093,8 @@ public final class ConfigurationAsyncClient {
      * <!-- src_embed com.azure.data.appconfiguration.configurationasyncclient.listConfigurationSettingsForSnapshotMaxOverload -->
      * <pre>
      * String snapshotName = &quot;&#123;snapshotName&#125;&quot;;
-     * List&lt;SettingFields&gt; fields = Arrays.asList&#40;SettingFields.KEY&#41;;
-     *
-     * client.listConfigurationSettingsForSnapshot&#40;snapshotName, fields&#41;
+     * SettingSelector selector = new SettingSelector&#40;&#41;.setFields&#40;SettingFields.KEY&#41;;
+     * client.listConfigurationSettingsForSnapshot&#40;snapshotName, selector&#41;
      *     .subscribe&#40;setting -&gt;
      *         System.out.printf&#40;&quot;Key: %s, Value: %s&quot;, setting.getKey&#40;&#41;, setting.getValue&#40;&#41;&#41;&#41;;
      * </pre>
@@ -1103,30 +1102,27 @@ public final class ConfigurationAsyncClient {
      *
      * @param snapshotName Optional. A filter used get {@link ConfigurationSetting}s for a snapshot. The value should
      * be the name of the snapshot.
-     * @param fields Optional. The fields to select for the query response. If none are set, the service will return the
-     * ConfigurationSettings with a default set of properties.
+     * @param selector Optional. Selector to filter configuration setting results from the service.
      * @return A Flux of ConfigurationSettings that matches the {@code selector}. If no options were provided, the Flux
      * contains all of the current settings in the service.
      * @throws HttpResponseException If a client or service error occurs, such as a 404, 409, 429 or 500.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
     public PagedFlux<ConfigurationSetting> listConfigurationSettingsForSnapshot(String snapshotName,
-                                                                                List<SettingFields> fields) {
+                                                                                SettingSelector selector) {
+        String acceptDateTime = selector == null ? null : selector.getAcceptDateTime();
         return new PagedFlux<>(
             () -> withContext(
                 context -> serviceClient.getKeyValuesSinglePageAsync(
-                        null,
-                        null,
-                        null,
-                        null,
-                        fields,
+                    null, null, null, acceptDateTime,
+                        selector == null ? null : toSettingFieldsList(selector.getFields()),
                         snapshotName,
                         addTracingNamespace(context))
                     .map(pagedResponse -> toConfigurationSettingWithPagedResponse(pagedResponse))),
             nextLink -> withContext(
                 context -> serviceClient.getKeyValuesNextSinglePageAsync(
                         nextLink,
-                        null,
+                        acceptDateTime,
                         addTracingNamespace(context))
                     .map(pagedResponse -> toConfigurationSettingWithPagedResponse(pagedResponse)))
         );
@@ -1265,8 +1261,8 @@ public final class ConfigurationAsyncClient {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     public Mono<Response<ConfigurationSettingsSnapshot>> getSnapshotWithResponse(String name,
-                                                                                List<SnapshotFields> fields) {
-        return serviceClient.getSnapshotWithResponseAsync(name, null, null, fields, Context.NONE)
+        Iterable<SnapshotFields> fields) {
+        return serviceClient.getSnapshotWithResponseAsync(name, null, null, iterableToList(fields), Context.NONE)
             .map(response -> new SimpleResponse<>(response, response.getValue()));
     }
 
@@ -1405,7 +1401,7 @@ public final class ConfigurationAsyncClient {
                     context -> serviceClient.getSnapshotsSinglePageAsync(
                         selector == null ? null : selector.getName(),
                         null,
-                        null,
+                        selector == null ? null : iterableToList(selector.getFields()),
                         selector == null ? null : iterableToList(selector.getSnapshotStatus()),
                         addTracingNamespace(context))),
                 nextLink -> withContext(